+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-localization` will be documented in this file
 
+## 1.1.0 - 2021-02-16
+
+- Added Facade `Localization`
+- Added Exceptions for checking `configHasLocales` and `LocaleInLocales`
+
 ## 1.0.0 - 2021-02-16
 
 - Initial release